requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=1.4
pyodbc
pydantic
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pydantic[email]
reportlab
